Update the filename
Rename the scripts: all xpm files start with '_'.

Move the help/history/dir/stack panels and debugger toolbar to bsm folder.

Update the ctx_start to support 'double' type duration.

+++ bsm/simengine.py
@@ -83,7 +83,7 @@ class SimEngine(object):
         PTFILE = POINTER(SimTraceFile)
         PTOBJ = POINTER(SimObj)
         PTBUF = POINTER(SimTraceBuf)
-        interfaces = [['start', '', (c_int, c_int), None],
+        interfaces = [['start', '', (c_double, c_int), None],
                       ['stop', '', None, None],
                       ['time', '', None, c_double],
                       ['time_sec', '', (c_double, c_int), c_double],
